ex_vantage v0.0.1 AirVantage.API

Utilities for interacting with the AirVantage API.

Types

body()

body() :: String.t()

endpoint()

endpoint() :: String.t()

headers()

headers() :: %{required(String.t()) => String.t()} | %{}

method()

method() :: :get | :post | :put | :delete

params()

params() :: map()

Functions

request(params, method, endpoint, headers, opts)

request(params(), method(), endpoint(), headers(), list()) ::
  {:ok, map()} | {:error, AirVantage.Error.t()}

A low level function which performs any request to AirVantage API. This function could actually be used in order to make request to unsupported endpoints.