ex_vantage v0.0.1 AirVantage.Operations.System

The system is the core representation used in AirVantage to define and interact with a real system. A system is composed of:

a gateway: the physical module enabling the connectivity of the system

a subscription: a subscription is the configuration of the connectivity defined with the operator

one or several applications: an application defines a piece of logic executing on the system. It can be a firmware, a software, etc.

Functions

find(fields, gateway)

find(String.t(), String.t()) :: {:ok, map()} | {:error, AirVantage.Error.t()}

Returns a paginated list of systems with their complete details. It is possible to filter out the result list using criteria parameters. Though system creation date is not publicly exposed, the default sort order is based on the creation date.