ex_vantage v0.0.1 AirVantage.Operations.SystemDeviceManagement

Device managament related operations

Functions

wake_up(uids)

wake_up(list()) :: {:ok, map()} | {:error, AirVantage.Error.t()}

Wakes up a communication application deployed on a selection of systems in order to force this application to communicate with the Operating Portal.