ex_vantage v0.0.1 AirVantage.Request

A module to build and execute requests to AirVantage API. The request does not happen until its passed to make_request/1. At a minimum, a request must have the endpoint and method specified to be valid.

Types

t()

t() :: %AirVantage.Request{
  endpoint: AirVantage.API.endpoint() | nil,
  headers: AirVantage.API.headers() | nil,
  method: AirVantage.API.method() | nil,
  opts: Keyword.t(),
  params: map()
}

Functions

make_request(request)

make_request(t()) :: {:ok, map()} | {:error, AirVantage.Error.t()}

Executes the request and returns the response.

new_request(headers \\ %{"Content-Type" => "application/json"})

new_request(map() | nil) :: t()

Creates a new request.

put_endpoint(request, endpoint)

put_endpoint(t(), String.t()) :: t()

Specifies an endpoint for the request.

put_method(request, method)

put_method(t(), AirVantage.API.method()) :: t()

Specifies a method to use for the request. Accepts any of the standard HTTP methods as atoms, that is :get, :post, :put or :delete.

put_param(request, key, value)

put_param(t(), atom(), any()) :: t()

Specify a single param to be included in the request.

put_params(request, new_params)

put_params(t(), map()) :: t()

Specifies the parameters to be used for the request. If the request is a :post request, these are encoded in the request body. Otherwise, they are encoded in the URL. Calling this function multiple times will merge, not replace, the params currently specified.